+++ src/libwbfs/wiidisc.h
@@ -141,7 +141,7 @@ typedef struct wiidisc_t
 
 extern const char * wd_partition_name[];
 const char * wd_get_partition_name ( u32 ptype, const char * result_if_not_found );
-char * wd_print_partition_name ( char * buf, u32 buf_size, u32 ptype, int print_num );
+char * wd_print_partition_name ( char * buf, size_t buf_size, u32 ptype, int print_num );
 
 const u8 * wd_get_common_key();
 const u8 * wd_set_common_key ( const u8 * new_key );
